worker/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-analysis-worker"
version = "0.1.0"
description = "Chess.com game analysis worker with Stockfish integration"
authors = [{name = "Chess Analysis Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-chess==1.999",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    "psycopg2-binary==2.9.9",
    "stockfish==3.28.0",
]

[project.scripts]
ingest = "ingest:main"
seed = "seed_one_month:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
